requirements.txt
boto3>=1.34.0
langchain>=0.3.0
langchain-aws>=0.1.0
langchain-community>=0.3.0
langgraph>=0.1.0
faiss-cpu>=1.8.0
streamlit>=1.35.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
faker>=25.0.0
yfinance>=0.2.40
pandas>=2.0.0
"numpy<2"
aws-cdk-lib>=2.140.0
constructs>=10.0.0
